191111 | Authentication_회원가입, 로그인, 로그아웃, login_required

+++ 05_django/04_django_form/articles/views.py
@@ -1,4 +1,5 @@
 from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from IPython import embed
 from .models import Article
@@ -7,11 +8,12 @@ from .forms import ArticleForm
 from .forms import CommentForm
 
 def index(request):
+    # embed()
     articles = Article.objects.all()
     context = {'articles':articles,}
     return render(request,'articles/index.html', context)
 
-
+@login_required # 데코레이트 지정
 def create(request):
     
     if request.method == 'POST':
@@ -52,10 +54,13 @@ def detail(request, article_pk):
         } 
     return render(request, 'articles/detail.html', context)
 
+# 동시에 쓰고 싶은 경우 @login_required를 내부에서 처리해준다.
+# @login_required
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article,pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article,pk=article_pk)
+        article.delete()
     return redirect('articles:index')
     
     # if request.method == 'POST':    
@@ -64,7 +69,7 @@ def delete(request, article_pk):
     # else:
     #     return redirect('articles:detail', article.pk)
 
-
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article,pk=article_pk)
     
@@ -96,14 +101,15 @@ def update(request, article_pk):
 @require_POST
 def comments_create(request, article_pk):
     article = get_object_or_404(Article,pk=article_pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        # save 메서드 -> 선택 인자 : (기본값) commit=True -> DB에 저장
-        # DB에 바로 저장되는 것을 막아준다. -> commit=False
-        comment = comment_form.save(commit=False) 
-        comment.article = article
-        comment.save()
-    return redirect('articles:detail', article.pk)
+    if request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            # save 메서드 -> 선택 인자 : (기본값) commit=True -> DB에 저장
+            # DB에 바로 저장되는 것을 막아준다. -> commit=False
+            comment = comment_form.save(commit=False) 
+            comment.article = article
+            comment.save()
+        return redirect('articles:detail', article.pk)
 
     # if request.method == 'POST':
     #     comment_form = CommentForm(request.POST)
@@ -119,8 +125,9 @@ def comments_create(request, article_pk):
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment = Comment.objects.get(pk=comment_pk)
+        comment.delete()
     return redirect('articles:detail', article_pk)
 
     # if request.method == 'POST':
